backend/requirements.txt
# Web Framework and API
Flask==3.0.3
Flask-Cors==4.0.1
fastapi==0.115.0
uvicorn==0.31.0

# Data Processing
pandas==2.2.2
openpyxl==3.1.5

# Database
mysql-connector-python==8.3.0
SQLAlchemy==2.0.23

# HTTP and Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2

# Google Generative AI
google-ai-generativelanguage==0.10.0

# Machine Learning and NLP
spacy==3.8.2
transformers==4.45.2
torch==2.4.1
accelerate==0.34.2

# Data Validation
pydantic==2.9.2

# Utilities
python-dotenv==1.0.1
retry==0.9.2

# Server Deployment
gunicorn==23.0.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0

# spaCy Language Model
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
